+++ lib/src/controller/group_controller.dart
@@ -70,4 +70,69 @@ class GroupController implements BaseController {
   void enableAll() {
     _widgetState.enableAll();
   }
+
+  @override
+  void select<k>(k value) {
+    assert(!(value is List),
+        "you should use `selectItems` instead of select is only for one value");
+    assert(_widgetState.values.contains(value),
+        "you cannot select  item that doesn't exist");
+
+    switch (isMultipleSelection) {
+      case true:
+        _widgetState.selectValues(List.filled(1, value) as List<k>);
+        break;
+      case false:
+        final index = _widgetState.values.indexOf(value);
+        _widgetState.notifierItems[index].value =
+            _widgetState.notifierItems[index].value.copy(
+          checked: true,
+        );
+        if (_widgetState.selectedValue.value != null) {
+          final indexOld =
+              _widgetState.values.indexOf(_widgetState.selectedValue.value);
+          if (indexOld != -1) {
+            _widgetState.notifierItems[indexOld].value =
+                _widgetState.notifierItems[indexOld].value.copy(
+              checked: false,
+            );
+          }
+        }
+        _widgetState.selectedValue.value = value;
+        break;
+    }
+  }
+
+  @override
+  void selectAll() {
+    assert(
+      isMultipleSelection,
+      "you cannot use selectAll in single selection group checkbox",
+    );
+    _widgetState.selectValues(_widgetState.values);
+  }
+
+  @override
+  void selectItems<k>(List<k> values) {
+    assert(isMultipleSelection,
+        "you cannot select multiple items in single selection group");
+    if (values.length == 1) {
+      select(values.first);
+    }
+    _widgetState.selectValues(values);
+  }
+
+  @override
+  void deselectValues<k>(List<k> values) {
+    assert(isMultipleSelection,
+        "you cannot deselect multiple items in single selection group");
+    _widgetState.deselectValues(values);
+  }
+
+  @override
+  void deselectAll() {
+    assert(isMultipleSelection,
+        "you cannot deselect all items in single selection group");
+    _widgetState.deselectValues(_widgetState.values);
+  }
 }