+++ test/checkbox_grouped_test.dart
@@ -215,4 +215,210 @@ void main() {
     await tester.pump();
     expect(controller.selectedItem, 3);
   });
+
+  testWidgets("test One Select GroupController single selection",
+      (tester) async {
+    GroupController controller = GroupController();
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(
+        body: SimpleGroupedCheckbox<int>(
+          controller: controller,
+          itemsTitle: List.generate(10, (index) => "${index + 1}"),
+          values: List.generate(10, (index) => index + 1),
+          activeColor: Colors.red,
+          checkFirstElement: true,
+          isLeading: true,
+        ),
+      ),
+    ));
+    //await tester.pump(Duration(seconds: 5));
+    await tester.pump();
+    //await tester.tap(find.byType(RadioListTile).first);
+    //var rb0 = tester.widget(find.byElementType(RadioListTile).first) as RadioListTile<int>;
+    expect(controller.selectedItem, 1);
+    await tester.pump();
+    controller.select(3);
+    await tester.pump();
+    expect(controller.selectedItem, 3);
+    controller.select(4);
+    await tester.pump();
+    expect(controller.selectedItem, 4);
+
+    /// test no accepting list
+    expect(() => controller.select([4, 3]), throwsA(isA<AssertionError>()));
+  });
+
+  testWidgets("test One Select GroupController multiple selection",
+      (tester) async {
+    GroupController controller = GroupController(isMultipleSelection: true);
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(
+        body: SimpleGroupedCheckbox<int>(
+          controller: controller,
+          itemsTitle: List.generate(10, (index) => "${index + 1}"),
+          values: List.generate(10, (index) => index + 1),
+          activeColor: Colors.red,
+          checkFirstElement: true,
+          isLeading: true,
+        ),
+      ),
+    ));
+    //await tester.pump(Duration(seconds: 5));
+    await tester.pump();
+    //await tester.tap(find.byType(RadioListTile).first);
+    //var rb0 = tester.widget(find.byElementType(RadioListTile).first) as RadioListTile<int>;
+    expect(controller.selectedItem, [1]);
+    await tester.pump();
+    controller.select(3);
+    await tester.pump();
+    expect(controller.selectedItem, [1, 3]);
+    controller.select(4);
+    await tester.pump();
+    expect(controller.selectedItem, [1, 3, 4]);
+  });
+
+  testWidgets("test select all in multiple selection  SimpleGroupedCheckbox",
+      (tester) async {
+    GroupController controller =
+        GroupController(isMultipleSelection: true, initSelectedItem: [1, 2]);
+    final values = List.generate(10, (index) => index);
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(
+        body: SimpleGroupedCheckbox<int>(
+          controller: controller,
+          itemsTitle: List.generate(10, (index) => "$index"),
+          values: values,
+          activeColor: Colors.green,
+          checkFirstElement: false,
+          onItemSelected: (data) {
+            print(data);
+          },
+          isExpandableTitle: false,
+        ),
+      ),
+    ));
+    await tester.pump();
+    //tester.allElements;
+    expect(controller.selectedItem, [1, 2]);
+    controller.selectAll();
+    await tester.pump();
+    final controllerValues = List.from(controller.selectedItem);
+    controllerValues.sort();
+    expect(controllerValues, values);
+  });
+
+  testWidgets(
+      "test select some values in multiple selection  SimpleGroupedCheckbox",
+      (tester) async {
+    GroupController controller =
+        GroupController(isMultipleSelection: true, initSelectedItem: [1, 2]);
+    final values = List.generate(10, (index) => index);
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(
+        body: SimpleGroupedCheckbox<int>(
+          controller: controller,
+          itemsTitle: List.generate(10, (index) => "$index"),
+          values: values,
+          activeColor: Colors.green,
+          checkFirstElement: false,
+          onItemSelected: (data) {
+            print(data);
+          },
+          isExpandableTitle: false,
+        ),
+      ),
+    ));
+    await tester.pump();
+    //tester.allElements;
+    expect(controller.selectedItem, [1, 2]);
+    controller.selectItems([3, 5]);
+    await tester.pump();
+    final controllerValues = List.from(controller.selectedItem);
+    controllerValues.sort();
+    expect(controllerValues, [1, 2, 3, 5]);
+  });
+  testWidgets(
+      "test deselect some values in multiple selection  SimpleGroupedCheckbox",
+      (tester) async {
+    GroupController controller =
+        GroupController(isMultipleSelection: true, initSelectedItem: [1, 2]);
+    final values = List.generate(10, (index) => index);
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(
+        body: SimpleGroupedCheckbox<int>(
+          controller: controller,
+          itemsTitle: List.generate(10, (index) => "$index"),
+          values: values,
+          activeColor: Colors.green,
+          checkFirstElement: false,
+          onItemSelected: (data) {
+            print(data);
+          },
+          isExpandableTitle: false,
+        ),
+      ),
+    ));
+    await tester.pump();
+    //tester.allElements;
+    expect(controller.selectedItem, [1, 2]);
+    controller.deselectValues([2]);
+    await tester.pump();
+    expect(controller.selectedItem, [1]);
+  });
+
+  testWidgets("test deselect all in multiple selection  SimpleGroupedCheckbox",
+      (tester) async {
+    GroupController controller =
+        GroupController(isMultipleSelection: true, initSelectedItem: [1, 2]);
+    final values = List.generate(10, (index) => index);
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(
+        body: SimpleGroupedCheckbox<int>(
+          controller: controller,
+          itemsTitle: List.generate(10, (index) => "$index"),
+          values: values,
+          activeColor: Colors.green,
+          checkFirstElement: false,
+          onItemSelected: (data) {
+            print(data);
+          },
+          isExpandableTitle: false,
+        ),
+      ),
+    ));
+    await tester.pump();
+    expect(controller.selectedItem, [1, 2]);
+    controller.deselectAll();
+    await tester.pump();
+    expect(controller.selectedItem, []);
+  });
+
+  testWidgets("test deselect all in one selection  SimpleGroupedCheckbox",
+      (tester) async {
+    GroupController controller = GroupController(
+      isMultipleSelection: false,
+    );
+    final values = List.generate(10, (index) => index);
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(
+        body: SimpleGroupedCheckbox<int>(
+          controller: controller,
+          itemsTitle: List.generate(10, (index) => "$index"),
+          values: values,
+          activeColor: Colors.green,
+          checkFirstElement: false,
+          onItemSelected: (data) {
+            print(data);
+          },
+          isExpandableTitle: false,
+        ),
+      ),
+    ));
+    await tester.pump();
+    controller.select(1);
+    await tester.pump();
+    expect(controller.selectedItem, 1);
+    await tester.pump();
+    expect(() => controller.deselectAll(), throwsA(isA<AssertionError>()));
+  });
 }