+++ test/custom_grouped_checkbox_test.dart
@@ -14,7 +14,10 @@ void main() {
               return CustomGroupedCheckbox<int>(
                 controller: controller,
                 itemBuilder: (ctx, index, v, isDisabled) {
-                  return Text("$index");
+                  return Text(
+                    "$index",
+                    key: Key("$index"),
+                  );
                 },
                 values: List<int>.generate(10, (i) => i),
               );
@@ -23,11 +26,13 @@ void main() {
         ),
       ),
     );
-    await tester.pump();
-    await tester.tap(find.byType(Text).at(1));
-    await tester.pump();
+    await tester.pumpAndSettle();
+
+    await tester.tap(find.byKey(Key("1")), warnIfMissed: false);
+    await tester.pumpAndSettle();
     expect(controller.selectedItem, 1);
-    await tester.tap(find.byType(Text).at(2));
+
+    await tester.tap(find.byKey(Key("2")), warnIfMissed: false);
     await tester.pump();
     expect(controller.selectedItem, 2);
   });
@@ -54,10 +59,10 @@ void main() {
     );
 
     await tester.pump();
-    await tester.tap(find.byType(Text).at(4));
+    await tester.tap(find.byType(Text).at(4), warnIfMissed: false);
     await tester.pump();
     expect(controller.selectedItem, [5]);
-    await tester.tap(find.byType(Text).at(5));
+    await tester.tap(find.byType(Text).at(5), warnIfMissed: false);
     await tester.pump();
     expect(controller.selectedItem, [5, 6]);
   });
@@ -75,6 +80,7 @@ void main() {
                 controller: controller,
                 itemBuilder: (ctx, index, v, isDisabled) {
                   return Container(
+                    key: Key("$index"),
                     color: isDisabled == true ? Colors.grey : null,
                     padding: EdgeInsets.all(5.0),
                     child: Text("$index"),
@@ -89,15 +95,15 @@ void main() {
     );
 
     await tester.pump();
-    await tester.tap(find.byType(Text).at(4));
+    await tester.tap(find.byType(Text).at(4), warnIfMissed: false);
     await tester.pump();
     expect(controller.selectedItem, [User("name5")]);
     controller.disabledItems([User("name6")]);
-    await tester.tap(find.byType(Text).at(5));
+    await tester.tap(find.byType(Text).at(5), warnIfMissed: false);
     await tester.pump();
     expect(controller.selectedItem, [User("name5")]);
     controller.enabledItems([User("name6")]);
-    await tester.tap(find.byType(Text).at(5));
+    await tester.tap(find.byType(Text).at(5), warnIfMissed: false);
     await tester.pump();
     expect(controller.selectedItem, [User("name5"), User("name6")]);
   });